requirements.txt
# Core
torch>=2.5.0
numpy>=1.26.0

# Data & Tokenization
datasets>=3.0.0
tokenizers>=0.20.0
tqdm>=4.66.0

# Tool-calling & Data Generation
openai>=1.50.0
python-dotenv>=1.0.0
ddgs>=9.0.0

# Config
PyYAML>=6.0.2

# Testing
pytest>=7.4.0
